% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-entropy.R
\name{cross_entropy_control}
\alias{cross_entropy_control}
\title{Hyperparmaeters for the cross-entropy models}
\usage{
cross_entropy_control(
  starts = 10,
  learning_rate = 0.05,
  maximum_time = 3600,
  iterations = 500,
  early_stopping = 25,
  thresholds = 100,
  split_percentiles = FALSE,
  samples = 100,
  elite_samples = 5,
  threads = parallel::detectCores() - 1
)
}
\arguments{
\item{starts}{Number of fast-and-frugal trees or tallying models generated with different seeds. The best fitting model will be return. The default is 10.}

\item{learning_rate}{The default is 0.1.}

\item{maximum_time}{Maximum training time (in seconds) to obtain the final model.}

\item{iterations}{The maximum number of iterations. Optimization may stop early if the model has converged. The default is 500.}

\item{early_stopping}{Training stops early if the performance does not improve after this many iterations in a row. The default is 25.}

\item{thresholds}{The maximum number of numeric thresholds that are tested for a numeric feature. The default is 100.}

\item{split_percentiles}{When TRUE, the candidate thresholds at which features can be split are the percentiles of the features' distribution. When FALSE the candidate values are obtained by dividing the feature's values in equidistant bins. The default value is FALSE.}

\item{samples}{Number of models created in each iteration. The default is 100.}

\item{elite_samples}{The number of best-performing models that are used to update the parameter distribution. The default is 10.}

\item{threads}{Number of CPU kernels used. The default is the number of available cores minus 1.}
}
\description{
\code{cross_entropy_control} Contains a list of parameters that control the nuances of the cross-entropy optimization procedure.
}
