% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\docType{class}
\name{fftreeModel-class}
\alias{fftreeModel-class}
\title{class for fast-and-frugal-tree models}
\description{
An S4 class to represent a fast-and-frugal tree model
}
\section{Slots}{

\describe{
\item{\code{call}}{An image of the call that produced the object.}

\item{\code{formula}}{\link[stats]{formula} Object of the model.}

\item{\code{parameters}}{Parameters used to train the fast-and-frugal tree.}

\item{\code{performance}}{List showing fitting and cross-validated performance.}

\item{\code{class_labels}}{A vector of length 2 containing the class labels. The second entry is referred to as the positive class.}

\item{\code{weights}}{A numeric vector of length 2. The first entry denotes the weight of instances in the negative class, the second entry the weight of instances in the positive class.}

\item{\code{prior}}{The proportion of objects in the positive class in the training set.}

\item{\code{tree}}{Representation of the tree.}

\item{\code{training_data}}{Data that was used to train the model}
}}

