% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver.R
\docType{data}
\name{liver}
\alias{liver}
\title{Liver patients}
\format{
A data frame with 579 rows and 11 variables:
\describe{
  \item{diagnosis}{whether subject is liver patient (414) or not (165), labeled by experts}
  \item{age}{age of patient}
  \item{sex}{either \code{'female'} or \code{'male'}}
  \item{totalBilirubin}{total bilirubin concentration in the blood}
  \item{directBilirubin}{direct bilirubin concentration in the blood}
  \item{alkaline}{alkaline phosphotase concentration}
  \item{alamine}{alamine aminotransferase concentration}
  \item{aspartate}{aspartate aminotransferase concentration}
  \item{proteins}{protein concentration}
  \item{albumin}{albumin concentration}
  \item{albuminGlobulin}{ albumin globulin ratio}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/ILPD+(Indian+Liver+Patient+Dataset)}
}
\usage{
liver
}
\description{
A dataset containing containing 414 liver and 165 non liver patient records.
The data was collected in the northeast of Andhra Pradesh, India.
}
\references{
Dua, Dheeru, and Graff, Casey "UCI machine learning repository." (2017)

Ramana, Bendi Venkata, M. Surendra Prasad Babu, and N. B. Venkateswarlu. "A critical study of selected classification algorithms for liver disease diagnosis." International Journal Of Database Management Systems 3.2 (2011): 101-114.
}
\keyword{datasets}
