% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fft.R
\name{plot,fftreeModel-method}
\alias{plot,fftreeModel-method}
\title{plots fast-and-frugal tree}
\usage{
\S4method{plot}{fftreeModel}(x, probabilities = F, legend = T, ...)
}
\arguments{
\item{x}{An object of type \linkS4class{fftreeModel-class}}

\item{probabilities}{If \code{TRUE} probability estimates are shown for each leaf. If \code{FALSE}, predicted class labels are shown.}

\item{legend}{If \code{TRUE} legend is shown.}

\item{...}{optional parameters passed to low level function}
}
\description{
plots fast-and-frugal tree
}
