% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.predict.R
\name{predict,fftreeModel-method}
\alias{predict,fftreeModel-method}
\title{Predictions from a fftree objects}
\usage{
\S4method{predict}{fftreeModel}(object, newdata, type = "response", ...)
}
\arguments{
\item{object}{an S4 object of class \code{\link{fftreeModel-class}} created by \code{\link{fftree}} function}

\item{newdata}{a data frame or matrix containing new data}

\item{type}{character string denoting the type of output returned. \code{response} (default) returns the predicted class label, \code{metric} returns the classification performance and \code{frugality} returns the average number of nodes visisted until a decision was made.}

\item{...}{additional arguments that are only used if the function is called internally}
}
\description{
Return predicted responses or measures of performance from a fitted \code{\link{fftreeModel-class}} object
}
