% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.predict.R
\name{predict,tallyModel-method}
\alias{predict,tallyModel-method}
\title{Predictions from a fftr objects}
\usage{
\S4method{predict}{tallyModel}(object, newdata, type = "response")
}
\arguments{
\item{object}{an S4 object of class \linkS4class{tallyModel-class} created by \code{\link{tally}} function}

\item{newdata}{a data frame or matrix containing new data}

\item{type}{character string denoting the type of output returned. \code{response} (default) returns the predicted class label, \code{probability} returns a matrix of class probabilities, \code{metric} returns the classification performance.}
}
\description{
Return predicted responses or measures of performance from a fitted \linkS4class{tallyModel-class} object
}
