% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fft.R
\name{show,fftreeModel-method}
\alias{show,fftreeModel-method}
\title{prints fast-and-frugal tree}
\usage{
\S4method{show}{fftreeModel}(object)
}
\arguments{
\item{object}{An object of type \linkS4class{fftreeModel}}
}
\description{
prints fast-and-frugal tree
}
