% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{show,splits-method}
\alias{show,splits-method}
\title{prints splits}
\usage{
\S4method{show}{splits}(object)
}
\arguments{
\item{object}{An object of type \linkS4class{splits-splits}}
}
\description{
prints splits
}
