% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tally.R
\name{show,tallyModel-method}
\alias{show,tallyModel-method}
\title{prints tallying model}
\usage{
\S4method{show}{tallyModel}(object)
}
\arguments{
\item{object}{An object of type \linkS4class{tallyModel-class}}
}
\description{
prints tallying model
}
