% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{splitFeatures}
\alias{splitFeatures}
\alias{splitFeatures,data.frame-method}
\alias{splitFeatures,matrix-method}
\title{Splitting features}
\usage{
splitFeatures(data, formula = stats::as.formula(data), ...)

\S4method{splitFeatures}{data.frame}(
  data,
  formula = as.formula(data.frame(data)),
  splits = "gini",
  weights = c(1, 1),
  ...
)

\S4method{splitFeatures}{matrix}(
  data,
  formula = stats::as.formula(data.frame(data)),
  splits = "gini",
  weights = c(1, 1),
  ...
)
}
\arguments{
\item{data}{an object of class \code{\link[base]{data.frame}} or \code{\link[base]{matrix}}. The criterion can either be a factor with two levels or an integer \code{(0,1)}. The \emph{positive class} is the second factor level (\code{levels(data$criterion)[2]}), or \code{1} if the criterion is numeric.}

\item{formula}{\code{\link[stats]{formula}} (optional). If \code{formula} is not provided, the first column of the data argument is used as the response variable and all other columns as predictors.}

\item{...}{optional parameters passed to low level function}

\item{splits}{specifies the method used to find a splitting point on numeric and binary features
\itemize{
\item{gini (default)}
\item{entropy}
\item{median}
}}

\item{weights}{a numeric vector of length 2 (default: \code{c(1,1)}). The first entry specifies the weight of instances in the positive class, the second entry the weight of instances in the negative class.}
}
\value{
A \linkS4class{splits} object.
}
\description{
\code{splitFeatures} is used internally by \code{\link{fftree}} and \code{\link{tally}} to find split points on numeric and categorical features.
}
\examples{
data(liver)
splits <- splitFeatures(data = liver, formula = diagnosis~., splits = "median")
}
