% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{tally}
\alias{tally}
\alias{tally,data.frame-method}
\alias{tally,matrix-method}
\title{\code{tally} is used to fit a tallying model.}
\usage{
tally(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "basic",
  split_function = "gini",
  weights = c(1, 1),
  cv = FALSE,
  max_size = 6,
  cross_entropy_parameters = cross_entropy_control()
)

\S4method{tally}{data.frame}(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "basic",
  split_function = "gini",
  weights = c(1, 1),
  cv = FALSE,
  max_size = 6,
  cross_entropy_parameters = cross_entropy_control()
)

\S4method{tally}{matrix}(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "basic",
  split_function = "gini",
  weights = c(1, 1),
  cv = FALSE,
  max_size = 6,
  cross_entropy_parameters = cross_entropy_control()
)
}
\arguments{
\item{data}{an object of class \code{\link[base]{data.frame}} or \code{\link[base]{matrix}}. The criterion can either be a factor with two levels or an integer \code{(0,1)}. The \emph{positive class} is the second factor level (\code{levels(data$criterion)[2]}), or \code{1} if the criterion is numeric.}

\item{formula}{\code{\link[stats]{formula}} (optional). If \code{formula} is not provided, the first column of the data argument is used as the response variable and all other columns as predictors.}

\item{method}{type of induction method for the fast-and-frugal tree:
\itemize{
\item{regression}
\item{cross-entropy}
}}

\item{split_function}{Which function should be used to determine the splitting values on numeric features. This only applies to tallying models trained with the 'regression' method.
to the By default Gini entropy ('gini') is used. Other options are Shannono entropy ('entropy') and 'median'.}

\item{weights}{a numeric vector of length 2 (default: \code{c(1,1)}). The first entry specifies the weight of instances in the positive class, the second entry the weight of instances in the negative class.
(\emph{see examples}).}

\item{cv}{If \code{TRUE} 10-fold cross validation is used to estimate the predictive performance of the model.}

\item{max_size}{set maximum number of features that contribute to the tallying model (default:  6)}

\item{cross_entropy_parameters}{hyperparameters for the cross-entropy method. By default the output of the function \code{\link{cross_entropy_control}} is passed.}
}
\value{
A \linkS4class{tallyModel} object.
}
\description{
\code{tally} is used to fit a tallying model.
}
\examples{
data(liver)
model <- tally(data = liver, formula = diagnosis~.)
model

# weight instances by the inverse of the prior
# in this way both classes contribute equally when training the model
prior <- mean(ifelse(liver$diagnosis == "Liver disease", 1, 0))
weights <- c(1/prior, 1/(1-prior))
mod <- tally(data = liver, formula = diagnosis~., weights = weights)

}
