% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\docType{class}
\name{tallyModel-class}
\alias{tallyModel-class}
\title{class for tallying models}
\description{
An S4 class to represent a tallying tree model
}
\section{Slots}{

\describe{
\item{\code{call}}{An image of the call that produced the object.}

\item{\code{formula}}{\link[stats]{formula} object of the model.}

\item{\code{parameters}}{Parameters used to train the tallying model.}

\item{\code{performance}}{List showing fitting and cross-validated performance.}

\item{\code{class_labels}}{A vector of length 2 containing the class labels. The second entry is referred to as the positive class.}

\item{\code{weights}}{A numeric vector of length 2. The first entry denotes the weight of instances in the positive class, the second entry the weight of instances in the negative class.}

\item{\code{prior}}{The proportion of objects in the positive class in the training set.}

\item{\code{tally}}{Representation of the tallying model}

\item{\code{split_function}}{How numeric features are split when 'basic' method is used.}

\item{\code{training_data}}{Data that was used to train the model}
}}

