% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fftree}
\alias{fftree}
\alias{fftree,data.frame-method}
\alias{fftree,matrix-method}
\title{Fitting fast-and-frugal trees}
\usage{
fftree(
  data,
  formula = stats::as.formula(data),
  method = "greedy",
  split_function = "gini",
  max_depth = 6,
  weights = c(1, 1),
  pruning = FALSE,
  cv = FALSE,
  use_features_once = TRUE,
  cross_entropy_parameters = cross_entropy_control()
)

\S4method{fftree}{data.frame}(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "greedy",
  split_function = "gini",
  max_depth = 6,
  weights = c(1, 1),
  pruning = FALSE,
  cv = FALSE,
  use_features_once = TRUE,
  cross_entropy_parameters = cross_entropy_control()
)

\S4method{fftree}{matrix}(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "greedy",
  split_function = "gini",
  max_depth = 6,
  weights = c(1, 1),
  pruning = FALSE,
  cv = FALSE,
  use_features_once = TRUE,
  cross_entropy_parameters = cross_entropy_control()
)
}
\arguments{
\item{data}{an object of class \code{\link[base]{data.frame}} or \code{\link[base]{matrix}}. The criterion can either be a factor with two levels or an integer \code{(0,1)}. The \emph{positive class} is the second factor level (\code{levels(data$criterion)[2]}), or \code{1} if the criterion is numeric.}

\item{formula}{\code{\link[stats]{formula}} (optional). If \code{formula} is not provided, the first column of the data argument is used as the response variable and all other columns as predictors.}

\item{method}{type of induction method for the fast-and-frugal tree:
\itemize{
\item{greedy (default and recommended)}
\item{basic}
\item{cross-entropy}
}}

\item{split_function}{Which function should be used to determine the splitting values on numeric features. This only applies to fast-and-frugal trees with the 'basic' or 'greedy' method.

By default Gini entropy ('gini') is used. Other options are Shannon entropy ('entropy') and 'median'.}

\item{max_depth}{set maximum number of nodes of the fast-and-frugal tree (default: 6)}

\item{weights}{a numeric vector of length 2 (default: \code{c(1,1)}). The first entry specifies the weight of instances in the positive class, the second entry the weight of instances in the negative class.
(\emph{see examples}).}

\item{pruning}{by default pruning is not employed (\code{FALSE}). If the argument is set to \code{TRUE} the tree is pruned using cross-validation.}

\item{cv}{If \code{TRUE} 10-fold cross validation is used to estimate the predictive performance of the model.}

\item{use_features_once}{if \code{TRUE} an attribute is used only once in a tree. If \code{FALSE}, a feature may be split several times. Note that, by construction, the basic method can only use each feature once.}

\item{cross_entropy_parameters}{hyperparameters for the cross-entropy method. By default the output of the function \code{\link{cross_entropy_control}} is passed.}
}
\value{
A \linkS4class{fftreeModel} object.
}
\description{
\code{fftree} is used to fit fast-and-frugal trees.
}
\examples{
data(liver)
model <- fftree(data = liver, formula = diagnosis~., method = "greedy")
plot(model)
model

# weight instances by the inverse of the prior
# in this way both classes contribute equally when training the model
prior <- mean(ifelse(liver$diagnosis == "Liver disease", 1, 0))
weights <- c(1/prior, 1/(1-prior))
mod <- fftree(data = liver, formula = diagnosis~., weights = weights, method = "greedy")

}
