% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fft.R
\name{plot,fftreeModel-method}
\alias{plot,fftreeModel-method}
\title{plots fast-and-frugal tree}
\usage{
\S4method{plot}{fftreeModel}(
  x,
  proportions = FALSE,
  legend = TRUE,
  class_labels = FALSE,
  colors = c("brown3", "cornflowerblue"),
  ...
)
}
\arguments{
\item{x}{An object of type \linkS4class{fftreeModel-class}}

\item{proportions}{If \code{TRUE} the proportion of objects in the positive class are shown for each leaf. If \code{FALSE}, predicted class labels are shown.}

\item{legend}{If \code{TRUE}legend is shown.}

\item{class_labels}{If \code{TRUE} class labels are shown for each leaf.}

\item{colors}{Vector of length 2 to set the colors of the two classes. Default: \code{c("cornflowerblue""brown3")}.}

\item{...}{optional parameters passed to low level function}
}
\description{
plots fast-and-frugal tree
}
