% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\docType{class}
\name{splits-class}
\alias{splits-class}
\title{class for split object}
\description{
An S4 class to represent a split object
}
\section{Slots}{

\describe{
\item{\code{call}}{An image of the call that produced the object.}

\item{\code{type}}{Splitting function employed to split cues.}

\item{\code{formula}}{\link[stats]{formula} object of the model.}

\item{\code{class_labels}}{A vector of length 2 containing the class labels. The second entry is referred to as the positive class.}

\item{\code{weights}}{A numeric vector of length 2. The first entry denotes the weight of instances in the positive class, the second entry the weight of instances in the negative class.}

\item{\code{training_data}}{Data that was used to train the model}

\item{\code{splits}}{Representation of the splits}

\item{\code{prior}}{The proportion of objects in the positive class in the training set.}
}}

