% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{tally}
\alias{tally}
\alias{tally,data.frame-method}
\alias{tally,matrix-method}
\title{\code{tally} is used to fit a tallying model.}
\usage{
tally(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "basic",
  max_size = 6,
  split_function = "gini",
  weights = c(1, 1),
  cv = FALSE,
  cross_entropy_parameters = cross_entropy_control()
)

\S4method{tally}{data.frame}(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "basic",
  max_size = 6,
  split_function = "gini",
  weights = c(1, 1),
  cv = FALSE,
  cross_entropy_parameters = cross_entropy_control()
)

\S4method{tally}{matrix}(
  data,
  formula = stats::as.formula(data.frame(data)),
  method = "basic",
  max_size = 6,
  split_function = "gini",
  weights = c(1, 1),
  cv = FALSE,
  cross_entropy_parameters = cross_entropy_control()
)
}
\arguments{
\item{data}{An object of class \code{\link[base]{data.frame}} or \code{\link[base]{matrix}}. The response variable can either be a factor with two levels or an integer vector with values \code{0,1}.}

\item{formula}{\code{\link[stats]{formula}} (optional). If \code{formula} is not provided, the first column of the data argument is used as the response variable and all other columns as predictors.}

\item{method}{Type of induction method for the fast-and-frugal tree:
\itemize{
\item{basic} (default)
\item{cross-entropy}
}}

\item{max_size}{Maximum number of features that contribute to the tallying model (default: 6)}

\item{split_function}{Function should be used to determine the splitting values on numeric features. This only applies to tallying models trained with the 'basic' method. By default Gini entropy ('gini') is used. Other options are Shannono entropy ('entropy') and 'median'.}

\item{weights}{A numeric vector of length 2 (default: \code{c(1,1)}). The first entry specifies the weight of instances in the positive class, the second entry the weight of instances in the negative class.
(\emph{see examples}).}

\item{cv}{If \code{TRUE} 10-fold cross validation is used to estimate the predictive performance of the model. By default, pruning is not used.}

\item{cross_entropy_parameters}{Hyperparameters for the cross-entropy method. By default the output of the function \code{\link{cross_entropy_control}} is passed.}
}
\value{
A \linkS4class{tallyModel} object.
}
\description{
\code{tally} is used to fit a tallying model.
}
\examples{
data(liver)
model <- tally(data = liver, formula = diagnosis~.)
model

# weight instances by the inverse of the prior
# in this way both classes contribute equally when training the model
prior <- mean(ifelse(liver$diagnosis == "Liver disease", 1, 0))
weights <- c(1/prior, 1/(1-prior))
mod <- tally(data = liver, formula = diagnosis~., weights = weights)

}
